requirements.txt
Flask==2.3.3
PyCryptodome==3.18.0
Pillow==10.0.0
opencv-python==4.8.1.78
